split up the gui code into application.py and numberedtext.py
